Refactor command type and word
