+++ src/main/java/seedu/address/logic/parser/events/RemindEventCommandParser.java
@@ -7,7 +7,7 @@ import java.util.stream.Stream;
 
 import seedu.address.commons.core.index.Index;
 import seedu.address.commons.exceptions.IllegalValueException;
-import seedu.address.logic.commands.events.RemindEventCommand;
+import seedu.address.logic.commands.reminders.RemindEventCommand;
 import seedu.address.logic.parser.ArgumentMultimap;
 import seedu.address.logic.parser.ArgumentTokenizer;
 import seedu.address.logic.parser.Parser;
